requirements.txt
decouple==0.0.7
Django==5.1.7
django-cors-headers==4.3.1
fonttools==4.55.4
huggingface-hub==0.29.1
ipython==8.32.0
json5==0.10.0
langchain==0.3.22
langchain-cli==0.0.36
langchain-community==0.3.20
langchain-core==0.3.49
langchain-openai==0.3.11
langchain-text-splitters==0.3.7
langserve==0.3.1
langsmith==0.3.21
matplotlib==3.10.0
matplotlib-inline==0.1.7
mysqlclient==2.2.4
numpy==2.2.1
openai==1.70.0
opencv-python==4.10.0.84
openpyxl==3.1.5
pandas==2.2.3
pillow==10.4.0
psutil==6.0.0
py2neo==2021.2.4
scikit-image==0.25.2
scikit-learn==1.6.0
scipy==1.14.1
seaborn==0.13.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tiktoken==0.9.0
torch==2.5.1
torchaudio==0.12.1+cu113
torchvision==0.20.1
tqdm==4.67.1
ultralytics==8.1.0
websocket==0.2.1
websocket-client==1.8.0
websockets==14.2
